backend-codes/requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-environ>=0.11.2
django-allauth>=0.57.0
dj-rest-auth>=5.0.1
google-auth>=2.23.0
requests>=2.31.0
python-dotenv>=1.0.0
